fix(map): clean up map utilities -- move Leaflet function (prototype) to base-map; use base-map/util/ function rather than duplicate here; add fetch polyfill as a test for IE compatability
