+++ packages/base-map/src/index.js
@@ -6,6 +6,29 @@ import L from "leaflet";
 
 import callIfValid from "./util";
 
+/**
+ * panToOffset will allow you to pan the map and adjust for something like a floating
+ * left nav bar, or a page header with an offset center
+ *
+ * @note adapted from Peter's code: https://gist.github.com/missinglink/7620340
+ *
+ * @param latlng
+ * @param offsetX & offsetY: defaults to [0, 0] ([X, Y] pixel offsets center) a positive x
+ * offset to shift the center to the right, and a positive y offset to shift the center to the
+ * bottom. Negatives will move to the center point left and top.
+ * @param options: pan options https://leafletjs.com/reference.html#pan-options
+ * @return return value from a call to https://leafletjs.com/reference.html#map-panto
+ */
+L.Map.prototype.panToOffset = function(latlng, offsetX, offsetY, options) {
+  const x =
+    this.latLngToContainerPoint(latlng).x - (parseInt(offsetX, 10) || 0);
+  const y =
+    this.latLngToContainerPoint(latlng).y - (parseInt(offsetY, 10) || 0);
+  const point = this.containerPointToLatLng([x, y]);
+  /* eslint-disable-next-line no-underscore-dangle */
+  return this.setView(point, this._zoom, { pan: options });
+};
+
 // eslint-disable-next-line func-names
 L.Evented.addInitHook(function() {
   if (this) this.singleClickTimeout = null;
@@ -44,6 +67,8 @@ L.Evented.include({
     }
   }
 });
+
+
 
 /**
  * The base OpenTripPlanner map on which everything else is rendered.
